% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crop_rotation}
\alias{crop_rotation}
\title{Crop rotation}
\format{A data frame with 2 variables: \code{year} and \code{crop}.}
\usage{
crop_rotation
}
\description{
The type of crop in the watersheds by year.
}
\keyword{datasets}

