% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRIPS.R
\docType{package}
\name{STRIPS}
\alias{STRIPS}
\alias{STRIPS-package}
\title{STRIPS: A package for loading publically available STRIPS data.}
\description{
The STRIPS package provides a user-friendly mechanism to load all publically
availabe STRIPS data by simply installing and then loading this package. As
data are made public, this package will depend on those packages and thus
reinstalling the updated version of this package will allow access to those
data newly made public.
}
\section{STRIPS packages}{

  \code{\link[STRIPSArbuckle]{STRIPSArbuckle}}
  \code{\link[STRIPSHelmers]{STRIPSHelmers}}
  \code{\link[STRIPSHladik]{STRIPSHladik}}
  \code{\link[STRIPSKolka]{STRIPSKolka}}
  \code{\link[STRIPSLiebman]{STRIPSLiebman}}
  \code{\link[STRIPSMeta]{STRIPSMeta}}
  \code{\link[STRIPONeal]{STRIPSONeal}}
  \code{\link[STRIPSSchulte]{STRIPSSchulte}}
  \code{\link[STRIPSyield]{STRIPSyield}}
}
\seealso{
\code{\link[STRIPSMeta]{STRIPSMeta}}
}

