% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{watersheds}
\alias{watersheds}
\title{Watershed details}
\format{A data frame with 11 variables: \code{watershed}, \code{NAD83_X},
  \code{NAD83_Y}, \code{site}, \code{size_ha}, \code{slope_pct},
  \code{location_and_percentage_of_grass_filers}, \code{treatment}, and
  \code{prairie_pct}.}
\usage{
watersheds
}
\description{
Watershed meta data to be used for analyses.
}
\keyword{datasets}

